fix(files): arrange files structure
