+++ lib/msd_odata.rb
@@ -1,10 +1,10 @@
 require 'msd_odata/version'
 require 'msd_odata/util/logging'
+require 'msd_odata/util/url_helper'
 require 'msd_odata/client'
 require 'msd_odata/auth'
 require 'msd_odata/service'
 require 'msd_odata/entity'
-require 'msd_odata/url_helper'
 
 module MsdOdata
   class Error < StandardError; end
