+++ spec/lib/msd_odata/entity_spec.rb
