+++ spec/lib/msd_odata/service_spec.rb
