+++ spec/lib/msd_odata/util/url_helper_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 
-describe 'MsdOdata::UrlHelper' do
+describe 'MsdOdata::Util::UrlHelper' do
   describe '.entity_collection_url' do
     before(:each) do
       @entity = MsdOdata::Entity.new('JournalHeaders')
       @base_url = 'github.com'
-      @url_helper = MsdOdata::UrlHelper.new(@base_url, @entity)
+      @url_helper = MsdOdata::Util::UrlHelper.new(@base_url, @entity)
     end
 
     it 'builds url for entity collection' do
@@ -27,7 +27,7 @@ describe 'MsdOdata::UrlHelper' do
     it 'builds collection url with an empty query' do
       entity = MsdOdata::Entity.new('Invoices', {})
       base_url = 'google.com'
-      url_helper = MsdOdata::UrlHelper.new(base_url, entity)
+      url_helper = MsdOdata::Util::UrlHelper.new(base_url, entity)
 
       expected_url = "google.com/data/Invoices?"
       actual_url = url_helper.entity_collection_url(with_query: true)
@@ -39,7 +39,7 @@ describe 'MsdOdata::UrlHelper' do
     it 'builds url for entity' do
       entity = MsdOdata::Entity.new('Customers', { url_params: { 'CustomerId' => 'HS-120' } })
       base_url = 'github.com'
-      url_helper = MsdOdata::UrlHelper.new(base_url, entity)
+      url_helper = MsdOdata::Util::UrlHelper.new(base_url, entity)
 
       expected_url = "github.com/data/Customers(CustomerId='HS-120')"
 
@@ -51,7 +51,7 @@ describe 'MsdOdata::UrlHelper' do
     before(:each) do
       @entity = MsdOdata::Entity.new('JournalLines')
       @base_url = 'google.com'
-      @url_helper = MsdOdata::UrlHelper.new(@base_url, @entity)
+      @url_helper = MsdOdata::Util::UrlHelper.new(@base_url, @entity)
     end
 
     it 'builds a query with all options' do
